optimize temporary and permanent limits db writings and db reading (#82)
We do this slow data migration asynchronously in parallel to the server normal operations over the coming weeks. We force the migration when writing new data. In a few weeks, we will force the migration at startup of the server of all the remaining data that has not been migrated yet.

Signed-off-by: Etienne Homer <etiennehomer@gmail.com>
Co-authored-by: Etienne LESOT <etienne.lesot@rte-france.com>